Increase timeout for complex @media WASM test
This test processes 4 CSS rules including @font-face, @keyframes,
@media, and [block]. WP nightly CI needs more time for the WASM
compilation to complete and update the block attribute.
