+++ tests/at-rules/blueprint.json
@@ -1,0 +1,18 @@
+{
+	"$schema": "https://playground.wordpress.net/blueprint-schema.json",
+	"login": false,
+	"steps": [
+		{
+			"step": "defineWpConfigConsts",
+			"consts": { "WP_DEBUG": true }
+		},
+		{
+			"step": "activatePlugin",
+			"pluginPath": "pattern-css/pattern-css.php"
+		},
+		{
+			"step": "runPHP",
+			"code": "<?php require '/wordpress/wp-content/plugins/pattern-css/tests/setup.php'; ?>"
+		}
+	]
+}
